feat: Initial support for 32-bit platforms
lxmonika now builds on x86 and ARM32, with the target version lowered
to Win10 2004.

A lot of hacks are involved to get this unsupported version of WDK to
build on Visual Studio 2022. We rely on the latest `.props` and
`.targets` to do the job for us, just that the binaries and headers
are for an older version of Windows.

An attempt was made to further lower the bar to RS1, but it turned out
that we used too many Win10 2004+ functions so we will stop here for
now.

None of the built binaries have been tested.

+++ lxmonika/src/reality.cpp
@@ -151,11 +151,11 @@ RlpFileRead(
 
             INT64 oEnd = min(oStart + szLength, pFile->Length);
 
-            memcpy(pBuffer, pFile->Data + oStart, oEnd - oStart);
+            memcpy(pBuffer, pFile->Data + oStart, (SIZE_T)(oEnd - oStart));
 
             if (pOffset == NULL)
             {
-                pFile->Offset = oEnd;
+                pFile->Offset = (SIZE_T)oEnd;
             }
             else
             {
@@ -164,7 +164,7 @@ RlpFileRead(
 
             if (pBytesTransferred != NULL)
             {
-                *pBytesTransferred = oEnd - oStart;
+                *pBytesTransferred = (SIZE_T)(oEnd - oStart);
             }
 
             return 0;
@@ -523,7 +523,7 @@ RlpFileSeek(
     }
     else
     {
-        pFile->Offset = newOffset;
+        pFile->Offset = (SIZE_T)newOffset;
         if (pResultOffset != NULL)
         {
             *pResultOffset = newOffset;
